remove Span from rty
