+++ flux-middle/src/rty/conv.rs
@@ -63,7 +63,7 @@ pub(crate) fn conv_defn(defn: &fhir::Defn) -> rty::Defn {
     let mut name_map = NameMap::default();
     let sorts = name_map.conv_refined_by(&defn.args);
     let expr = Binders::new(name_map.conv_expr(&defn.expr, 1), sorts);
-    rty::Defn { name: defn.name, expr, span: defn.expr.span }
+    rty::Defn { name: defn.name, expr }
 }
 
 impl<'a, 'genv, 'tcx> ConvCtxt<'a, 'genv, 'tcx> {
